+++ schedule-base/src/main/kotlin/info/metadude/kotlin/library/schedule/v1/models/Day.kt
@@ -2,6 +2,7 @@ package info.metadude.kotlin.library.schedule.v1.models
 
 import info.metadude.kotlin.library.schedule.v1.serializers.LocalDateSerializer
 import info.metadude.kotlin.library.schedule.v1.serializers.OffsetDateTimeSerializer
+import info.metadude.kotlin.library.schedule.v1.serializers.RoomsSerializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import org.threeten.bp.LocalDate
@@ -18,5 +19,6 @@ data class Day(
     @SerialName("day_end")
     @Serializable(with = OffsetDateTimeSerializer::class)
     val dayEnd: OffsetDateTime,
+    @Serializable(with = RoomsSerializer::class)
     val rooms: Map<String, List<Event>>,
 )